requirements.txt
async-generator==1.10
attrs==22.2.0
beautifulsoup4==4.11.2
black==23.1.0
bs4==0.0.1
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
exceptiongroup==1.1.0
google-auth==2.16.0
google-auth-oauthlib==1.0.0
gspread==5.7.2
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
lxml==4.9.2
mypy-extensions==1.0.0
oauthlib==3.2.2
outcome==1.2.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
PySocks==1.7.1
python-dotenv==0.21.1
requests==2.28.2
requests-oauthlib==1.3.1
rsa==4.9
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
tomli==2.0.1
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typed-ast==1.5.4
typing_extensions==4.4.0
urllib3==1.26.14
webdriver-manager==3.8.5
wsproto==1.2.0
zipp==3.13.0
